APP_4_LANGCHAIN_AI_AGENT/requirements.txt
python-dotenv
langchain-core
langchain-google-genai
langchain
pydantic
todoist-api-python
tzdata